requirements.txt
flask
celery
redis
pillow
pytesseract